refactor: rename basic-parser-types to types
