+++ src/index.ts
@@ -30,7 +30,7 @@ export type {
     ImpalaSqlParserVisitor,
 } from './lib';
 
-export { EntityContextType } from './parser/common/basic-parser-types';
+export { EntityContextType } from './parser/common/types';
 
 export {
     /**
@@ -38,13 +38,9 @@ export {
      * It will be removed when the stable version is released.
      */
     EntityContextType as SyntaxContextType,
-} from './parser/common/basic-parser-types';
+} from './parser/common/types';
 
-export type {
-    CaretPosition,
-    Suggestions,
-    SyntaxSuggestion,
-} from './parser/common/basic-parser-types';
+export type { CaretPosition, Suggestions, SyntaxSuggestion } from './parser/common/types';
 
 export type { WordRange, TextSlice } from './parser/common/textAndWord';
