+++ test/parser/flink/benchmark/benchmark.md
@@ -1,12 +1,12 @@
 # FlinkSQL Benchmark
 | Name | Rows | Times | Total Time(ms) | Average Time(ms) |
 | ---- | ---- | ---- | ---- | ---- |
-| CreateTable | 100 | 1 | 193.17 | 193.17 |
-| CreateTable | 1000 | 1 | 70.31 | 70.31 |
-| CreateTable | 5000 | 1 | 201.97 | 201.97 |
-| SelectTable | 100 | 1 | 596.33 | 596.33 |
-| SelectTable | 1000 | 1 | 62.44 | 62.44 |
-| SelectTable | 5000 | 1 | 583.94 | 583.94 |
-| InsertTable | 100 | 1 | 8.52 | 8.52 |
-| InsertTable | 1000 | 1 | 47.36 | 47.36 |
-| InsertTable | 5000 | 1 | 268.04 | 268.04 |
+| CreateTable | 100 | 1 | 150.34 | 150.34 |
+| CreateTable | 1000 | 1 | 53.04 | 53.04 |
+| CreateTable | 5000 | 1 | 179.04 | 179.04 |
+| SelectTable | 100 | 1 | 460.25 | 460.25 |
+| SelectTable | 1000 | 1 | 46.24 | 46.24 |
+| SelectTable | 5000 | 1 | 505.28 | 505.28 |
+| InsertTable | 100 | 1 | 13.58 | 13.58 |
+| InsertTable | 1000 | 1 | 33.07 | 33.07 |
+| InsertTable | 5000 | 1 | 242.58 | 242.58 |
