+++ test/parser/spark/contextCollect/entityCollector.test.ts
@@ -3,7 +3,7 @@ import path from 'path';
 import { SparkSQL, SparkSqlSplitListener, SparkEntityCollector } from 'src/parser/spark';
 import { ParseTreeListener } from 'antlr4ng';
 import { SparkSqlParserListener } from 'src/lib/spark/SparkSqlParserListener';
-import { EntityContextType } from 'src/parser/common/basic-parser-types';
+import { EntityContextType } from 'src/parser/common/types';
 import { StmtContextType } from 'src/parser/common/entityCollector';
 
 const commonSql = fs.readFileSync(path.join(__dirname, 'fixtures', 'common.sql'), 'utf-8');
